requirements.txt
amqp==5.1.1
apt-xapian-index==0.49
apturl==0.5.2
argcomplete==3.1.2
asgiref==3.7.2
attrs==23.1.0
backoff==2.2.1
bcrypt==3.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
blinker==1.6.3
Brlapi==0.8.3
censys==2.2.7
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
command-not-found==0.3
cryptography==41.0.4
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.4.2
duplicity==0.8.21
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fasteners==0.14.1
filelock==3.12.4
Flask==3.0.0
Flask-Cors==4.0.0
Flask-WTF==1.2.1
future==0.18.3
gpg==1.16.0
h11==0.14.0
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kombu==5.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mongoengine==0.27.0
monotonic==1.6
more-itertools==8.10.0
netaddr==0.9.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
outcome==1.2.0
packaging==23.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
prompt-toolkit==3.0.39
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.5.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-nmap==0.7.1
python-whois==0.8.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
requests-file==1.5.1
rich==13.6.0
SecretStorage==3.3.1
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
ssh-import-id==5.11
systemd-python==234
termcolor==2.3.0
tldextract==3.6.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.6
usb-creator==0.3.7
vboxapi==1.0
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.8
Werkzeug==3.0.0
wsproto==1.2.0
WTForms==3.1.0
xdg==5
xkit==0.0.0
zipp==1.0.0
apt-xapian-index==0.49
apturl==0.5.2
argcomplete==3.1.2
asgiref==3.7.2
backoff==2.2.1
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.6.3
Brlapi==0.8.3
censys==2.2.7
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
command-not-found==0.3
cryptography==41.0.4
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.4.2
duplicity==0.8.21
fasteners==0.14.1
filelock==3.12.4
Flask==3.0.0
future==0.18.2
gpg==1.16.0
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
netaddr==0.9.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==23.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
requests-file==1.5.1
rich==13.6.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
ssh-import-id==5.11
systemd-python==234
termcolor==2.3.0
tldextract==3.6.0
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.6
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
Werkzeug==3.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
